@@ -5,9 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack -d -w --progress --colors",    
-    "start": "nodemon --harmony server/index",
-    "test-server": "anywhere -s -p 8001 -d ./build"
+    "build": "webpack -d -w --progress --colors",
+    "start": "nodemon server/entry"
   },
   "repository": {
     "type": "git",
@@ -20,34 +19,17 @@
   },
   "homepage": "https://github.com/javion25/chiji#readme",
   "dependencies": {
-    "bluebird": "^3.5.0",
-    "cheerio": "^1.0.0-rc.2",
     "koa": "^2.3.0",
     "koa-better-router": "^2.1.1",
-    "koa-static-server": "^1.3.2",
     "react": "^15.6.1",
     "react-dom": "^15.6.1"
   },
   "devDependencies": {
-    "anywhere": "^1.4.0",
-    "autoprefixer-loader": "^3.2.0",
     "babel-core": "^6.25.0",
     "babel-loader": "^7.1.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-regenerator": "^6.24.1",
     "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "browser-sync": "^2.18.13",
-    "css-loader": "^0.28.4",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
-    "html-webpack-plugin": "^2.30.1",
+    "babel-register": "^6.24.1",
     "nodemon": "^1.11.0",
-    "regenerator-runtime": "^0.10.5",
-    "stylus-loader": "^3.0.1",
-    "url-loader": "^0.5.9",
-    "vue-style-loader": "^3.0.1",
-    "webpack": "^3.4.1",
-    "webpack-node-externals": "^1.6.0"
+    "webpack": "^3.5.2"
   }
 }
